feat: after review - move SecondaryDbQueryService to search clients #29808
