+++ search/search-client/pom.xml
@@ -24,6 +24,11 @@
       <artifactId>camunda-security-core</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
+    </dependency>
+
   </dependencies>
 
 </project>
