+++ search/search-client/src/main/java/io/camunda/search/clients/SecondaryDbQueryService.java
@@ -5,14 +5,14 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.secondarydb;
+package io.camunda.search.clients;
 
 import io.camunda.search.filter.ProcessInstanceFilter;
 import io.camunda.search.page.SearchQueryPage;
 import io.camunda.search.query.SearchQueryResult;
 
 /**
- * This service provides a way to query the secondary database.
+ * This service provides a way to query the secondary database from zeebe engine.
  *
  * @apiNote This service is not intended to be used within a stream processor in the core engine, as
  *     it will block any threads/actors while the query is running. It should always be used in a
