+++ search/search-client/src/main/java/io/camunda/search/clients/impl/SecondaryDbQueryServiceImpl.java
@@ -5,15 +5,17 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.secondarydb;
+package io.camunda.search.clients.impl;
 
 import io.camunda.search.clients.ProcessInstanceSearchClient;
+import io.camunda.search.clients.SecondaryDbQueryService;
 import io.camunda.search.entities.ProcessInstanceEntity;
 import io.camunda.search.filter.ProcessInstanceFilter;
 import io.camunda.search.page.SearchQueryPage;
 import io.camunda.search.query.SearchQueryBuilders;
 import io.camunda.search.query.SearchQueryResult;
 
+/** This services uses the existing search client to query the secondary database. */
 public class SecondaryDbQueryServiceImpl implements SecondaryDbQueryService {
 
   private final ProcessInstanceSearchClient processInstanceSearchClient;
