+++ zeebe/engine/pom.xml
@@ -302,6 +302,7 @@
     <dependency>
       <groupId>io.camunda</groupId>
       <artifactId>camunda-search-domain</artifactId>
+      <scope>test</scope>
     </dependency>
   </dependencies>
