+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/clustering/ClusteringRule.java
@@ -381,6 +381,7 @@ public class ClusteringRule extends ExternalResource {
             new SecurityConfiguration(),
             null,
             null,
+            null,
             null);
 
     final Broker broker =
